feat: add chatbot
