+++ src/funcao_despedida.js
@@ -1,0 +1,9 @@
+function despedida() {
+    return (
+        "👋 *Encerrando a conversa...*\n\n" +
+        "Agradecemos por utilizar nossos serviços. Foi um prazer poder ajudá-lo(a) com suas dúvidas e necessidades. Caso precise de mais assistência no futuro, estaremos sempre à disposição para atendê-lo(a).\n\n" +
+        "Desejamos um excelente dia e sucesso em suas atividades. Até a próxima!"
+    );
+}
+
+module.exports = { despedida }
